Update return when no data for table view
Instead of return a blank when there is no data use the default return of 'None'.
This allows Symphony to display links if the field is set as the first entry and
thus in table view would have to be linked. When this was none no link was displayed
thus making the entry unreachable.